feat: finish audio note sharing
