+++ src/features/home/home-feed/components/audio-note.tsx
@@ -13,6 +13,7 @@ import { cn, getDateGroupLengths } from '@/lib/utils'
 import { trackEvent } from '@/utils/amplitude'
 import { formatTitle } from '@/utils/conversations'
 import { selectedAudioNoteAtom, sidePanelOpenAtom } from '@/atoms/side-panel'
+import { useAudioNotes } from '@/hooks/audio-notes'
 import { useCopyAudioShareLink, useGenerateAudioShareLink } from '@/hooks/share-link'
 import { Switch } from '@/components/ui/switch'
 import { Tooltip } from '@/components/ui/tooltip'
@@ -24,7 +25,6 @@ import { GroupedVirtualList, GroupHeaderRow } from '../components/grouped-virtua
 import { useInputFocus } from '../../chat-input/chat-input'
 import { currentListIndexAtom, isMountedAtom } from '../atoms'
 import { HOME_FEED_LIST_HEIGHT } from '../constants'
-import { useAudioNotes } from '../hooks'
 import { formatUpdatedAtDate } from '../utils'
 import { HomeFeedListItemLayout, HomeFeedListLayout, ListEmptyState, ListLoadingState } from './home-feed'
